Code reviewed. PR ready.
