+++ CMakeLists.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required(VERSION 3.22)
+
+project(LevelUp)
+
+set(CMAKE_CXX_STANDARD 20)
+
+add_executable(LevelUp
+        #task1.cpp
+        task2.cpp
+        )
